server/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
django-environ
django-filter
django-autoslug
django-countries
django-phonenumber-field
phonenumbers
isort
black
flake8
djangorestframework-simplejwt
PyJWT
djoser
google-cloud-storage
pytest-django
pytest-factoryboy
pytest-cov
Faker
redis
celery
django-celery-email
flower
argh
PyYAML
watchdog
